requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0
python-decouple>=3.8,<4.0
